Add server component
The server component serves a webpage
(http://<url>:8080?hostname=<individual name>) which can be opened by the client.
This webpage has an iframe and connects to the websocket server
(listening on port 3001).

Additionally the user is able to control the source of the clients
iframe with the admin component (http://<url>:8080/admin), which is
password protected.